Ensure pluralisations are loaded from rails-i18n
The `rails-i18n` gem provides standard pluralisations of various
locales. It activates these using a railtie, according to the
rails `available_locales` configuration setting.

If an app explicitly requires or loads `rails-i18n`, for example by
mentioning it in its gemfile, then all is well and the railtie will fire
properly.

However if an app uses this gem, which has a dependency on `rails-i18n`
and therefore does not explicitly require it, then the railtie will not
fire because the railtie definition will not have been required when
rails is running its initialization process.

By requiring rails-i18n in our main library file, we ensure that it the
railtie has been defined early on and that the railtie hooks will be in
place when rails initialization runs.

Prior to this I was seeing the following error when trying to run the
translation export rake task:

```
$ bundle exec rake translation:export:all[tmp/locales]
rake aborted!
No pluralization forms defined for locale 'ar'
```