Add updated list of plural forms
The original list of plural forms in govuk-app-config [1] is missing
some locales. This copies that list and adds some more languages we use
but don't have plural rules for. I've also updated some of the rules to
include 0 and 1 for 'one' plurals as per the rules defined on the
unicode website [2].

This list is intended to replace the one in govuk-app-config which we
use in the various Frontend apps, e.g collections [3]. This will occur
when we add the new features in Rails Translation Manager to the various
apps.

[1]: https://github.com/alphagov/govuk_app_config/blob/main/lib/govuk_app_config/govuk_i18n.rb
[2]: https://www.unicode.org/cldr/cldr-aux/charts/34/supplemental/language_plural_rules.html
[3]: https://github.com/alphagov/collections/blob/main/config/locales/plurals.rb