Update list of plural forms
The original list of plural rules are missing some locales that are
included in our frontend apps. This adds some more languages we use but
don't have plural rules for. I've also updated some of the rules to
include 0 and 1 for 'one' plurals as per the rules defined by unicode
[2].

This list is intended to replace the one in govuk-app-config [2] which we
will use in the various frontend apps, e.g Collections [3]. The
migration of the list will be completed in a future piece of work.

[1]: https://www.unicode.org/cldr/cldr-aux/charts/34/supplemental/language_plural_rules.html
[2]: https://github.com/alphagov/govuk_app_config/blob/main/lib/govuk_app_config/govuk_i18n.rb
[3]: https://github.com/alphagov/collections/blob/main/config/locales/plurals.rb