Add class to check for incorrect plural forms
IncompatiblePlurals checks for plural forms that have been incorrectly
added for all the languages in an application. It uses the native and
bespoke plural rules to detect inconsistencies in the locale files. E.g
English plurals are 'one' and 'other', if the locale keys just contained
'one' or any additional plural forms like 'few' then the report would
raise the discrepancy.

If a new language is added and there is no plural form in `rails-i18n`,
or the bespoke list (plurals.rb), then the code will output an error
requesting the user adds the correct rule.