Add the main checker class
LocaleChecker links up the other locale classes and generates a report
containing all our validation rules. It pulls the locales by searching
in the `locale_path` directory which is given by the calling app. E.g.
"...collection/config/locales/*.yml". The `available_locales` method
pulls the locales for the application and matches them against the files
in the locale path dir.

It will concatenate all locale keys for a langauge into one array to be
passed to the various classes.

The class outputs an error if the path provided doesn't match any eligible
YAML files which prevents false positives

The tests for this class act more like integration tests than 'unit'
tests as they test the code in the other checker classes, but I think
that is more useful than simply asserting class calls in this instance.