Add PluralForms class
PluralForms will pull all pluralization rules for the available
languages and put them into an array. The rules originate from
`rails-i18n` and the list we created to fill in the gaps at
`config/locales/plurals.rb`. This can then be used to validate plurals
against the locales plural form.