Use paths instead of `Rails.root`
This change allows us to be explicit about the path of the locale CSVs.

We've made this change primarily as `Rails.root` doesn't work as
expected with Govuk Publishing Components due to the root being
`spec/dummy`. However, it also has the benefit of being able to import
locales to any location.

As part of this we've also added tests for the import rake tasks.