Convert Importer test to spec
Converts the framework used for the Importer class from Minitest to
RSpec. This was also an opportunity to refactor the tests which removed
repetition and moved to using a real test CSV, instead of generating
them on the fly. As a result, the tests are easier to read and the line
count is reduced.