Update application root for app
We use `Rails.root` quite a lot in this Gem to figure out the calling
applications root directory. This is a nice means of working this out,
however for govuk_publishing_components this was being reported as
`govuk_publishing_component/spec/dummy` due to the Gem not being a true
Rails app and implementing it's own 'dummy' Rails app which was
overwriting the root dir.

This hack sets the true root directory for govuk_publishing_components
and is written so it should be future proof if it changes to something
other than `spec/dummy`.