Rename class
The class `TranslationHelper` was causing some strange errors to occur
in one the apps (Whitehall) that bundle Rails Translation Manager.

The error that was produced when running rake (in Whitehall):

```
TypeError: wrong argument type Class (expected Module)
/govuk/whitehall/test/test_helper.rb:300:in `<main>'
/govuk/whitehall/test/functional/admin/about_pages_controller_test.rb:1:in `<main>'
/root/.rbenv/versions/2.6.6/bin/bundle:23:in `load'
/root/.rbenv/versions/2.6.6/bin/bundle:23:in `<main>'
Tasks: TOP => test
```

This error message wasn't very clear, and so by stepping through the
code in RTM it appears the class `TranslationHelper` was causing the
issue. Not sure exactly why, but I assume it's because Rails already
defines `TranslationHelper` as a module [1].

As part of this I've renamed the class and moved it to where the rest of
the ruby scripts are defined.

[1]: https://api.rubyonrails.org/v5.1.7/classes/ActionView/Helpers/TranslationHelper.html