Make importing more resilient
In a number of locale files we received from the translators, there was
blank rows added (`,,`) to the CSV. This breaks the code as we don't
expect the row column `key` to be `nil`.

This commit removes rows which have blank keys and outputs the offending
rows to the console. This allows the import to run successfully while
giving the person running the import visibility of the rows that were
skipped.

Trello:
https://trello.com/c/Xz78vNR4/2757-import-newly-sourced-translations-into-locale-files-5