Add method to retrieve configurable root of locale files
This `locale_root` method has been added so that all of the references
to `Rails.root` have been replaced. My motivation for making this change
is spending an hour baffled by some strange behaviour in
govuk_publishing_components only to find it was a hack for this gem [1].

This change means that an environment variable can be set in any apps
that use this tool which don't follow common Rails conventions (or are
Rails engines).

[1]: https://github.com/alphagov/govuk_publishing_components/pull/2400/commits/1601d6a947a7d4dc5ca62ef422f719f1044a15e5