Use RailsTranslationManager.locale_root over Rails.root join
This relies on the top level helper method in RailsTranslationManager as
a way to access the directory with the Rails locale files within.

By using RailsTranslationManager.locale_root we can also configure the
path for any locale path we choose (via an env var) - it also has the
benefit of simplifying a bunch of lines of code here.