Consistently avoid outputting to stdout in tests
This makes use of the RSpec output.to_stdout for managing all the
writing to stdout.

Before this change:

```
➜  rails_translation_manager git:(26a6755) rspec
...Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.Invalid row: #<CSV::Row "key":nil "source":nil "translation":nil> for csv_path: spec/locales/importer/fr.csv
.....................................
Imported CSV from: path/to/import/fr.csv to spec/config/locales
..
Imported all CSVs from: spec/locales/importer to spec/config/locales
.......

Finished in 0.32614 seconds (files took 0.45576 seconds to load)
58 examples, 0 failures
```

After:

```
➜  rails_translation_manager git:(update-english-checker) rspec
..........................................................

Finished in 0.32423 seconds (files took 0.39752 seconds to load)
58 examples, 0 failures
```

Not all cases of applying this were particularly clean, the
RailsTranslationManager::LocaleChecker was a bit tricky to apply. I
concluded that the reason it's a bit strange is that the method that is
being tested both outputs to stdout and returns a boolean, which seemed
a bit of a dual purpose. I think it would be better a class if it had
separate methods to test and to output a result - which would also make
the testing neater.